Move set_backend by str and byte tests into public_api tests
